requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.25.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
